pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gmp-seeker"
version = "0.1.0"
description = "用于爬取 ISPE 指南文档信息的工具"
authors = [
    { name="Your Name", email="your@email.com" }
]
dependencies = [
    "Flask==2.3.2",
    "Flask-Compress>=1.13",
    "Flask-SQLAlchemy==3.0.5",
    "Flask-Login==0.6.3",
    "Flask-Admin==1.6.1",
    "Flask-WTF==1.1.1",
    "WTForms==3.0.1",
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
    "Werkzeug==2.3.6",
    "Jinja2==3.1.2",
    "python-dotenv>=0.9.9",
    "pypdf==3.15.4",
    "email-validator>=2.2.0",
    "openai>=1.97.1",
    "openpyxl>=3.1.5",
    "Markdown>=3.5",
    "bleach>=6.1.0",
    "boto3>=1.34.0",
]
requires-python = ">=3.8"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["logs*"]
